New version of handler supporting serverless version 1.5.0
