--- authorizer/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "authorizer",
-  "version": "1.0.0",
-  "description": "",
-  "main": "handler.js",
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "3scale": "^0.7.2",
-    "q": "^1.4.1",
-    "redis": "^2.5.3",
-    "redis-commands": "^1.1.0",
-    "request": "^2.71.0",
-    "then-redis": "^2.0.0",
-    "underscore": "*"
-  }
-}
