--- authorizer/s-function.json
@@ -1,39 +1,0 @@
-{
-  "name": "authorizer",
-  "runtime": "nodejs4.3",
-  "description": "Serverless Lambda function for project: sls-threeScale-Authorizer",
-  "customName": false,
-  "customRole": false,
-  "handler": "handler.handler",
-  "timeout": 6,
-  "memorySize": 1024,
-  "authorizer": {},
-  "custom": {
-    "excludePatterns": [],
-    "optimize": {
-      "exclude": [
-        "aws-sdk"
-      ]
-    }
-  },
-  "endpoints": [],
-  "events": [],
-  "environment": {
-    "SERVERLESS_PROJECT": "${project}",
-    "SERVERLESS_STAGE": "${stage}",
-    "SERVERLESS_REGION": "${region}",
-    "THREESCALE_PROVIDER_KEY": "YOUR_3SCALE_PROVIDER_KEY",
-    "THREESCALE_SERVICE_ID": "YOUR_3SCALE_SERVICE_ID",
-    "ELASTICACHE_ENDPOINT": "${elastiCacheEndpoint}",
-    "ELASTICACHE_PORT": 6379,
-    "SNS_TOPIC_ARN":"${snsTopic}"
-  },
-  "vpc": {
-    "securityGroupIds": [
-      "${vpcSecurityGroup}"
-    ],
-    "subnetIds": [
-      "${privateSubnetId1}","${privateSubnetId2}"
-    ]
-  }
-}
