+++ authorizer1.5.0/iamRoleStatements.json
@@ -1,0 +1,29 @@
+[
+  {
+    "Effect": "Allow",
+    "Action": [
+      "logs:CreateLogGroup",
+      "logs:CreateLogStream",
+      "logs:PutLogEvents"
+    ],
+    "Resource": "arn:aws:logs:${self:provider.region}:*:*"
+  },
+  {
+    "Effect": "Allow",
+    "Action": [
+      "ec2:CreateNetworkInterface",
+      "ec2:DescribeNetworkInterfaces",
+      "ec2:DeleteNetworkInterface"
+    ],
+    "Resource": "*"
+  },
+  {
+      "Effect": "Allow",
+      "Action": [
+          "sns:Publish"
+      ],
+      "Resource": [
+          {"Ref":"SNStopic"}
+      ]
+  }
+]
