+++ package.json
@@ -1,18 +1,18 @@
 {
-  "name": "sls-threeScale-Authorizer",
-  "version": "0.0.1",
-  "description": "A Serverless Project and its Serverless Plugin dependencies.",
-  "author": "me",
-  "license": "MIT",
-  "private": false,
-  "scripts":{
-    "postinstall":"./configure"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/"
-  },
+  "name": "awsThreeScale_Authorizer",
+  "version": "1.0.0",
+  "description": "Custom authorizer for 3scale API management solution integrated to AWS API Gateway",
+  "main": "handler.js",
+  "scripts": {},
+  "author": "Nicolas Grenié - @picsoung",
+  "license": "",
   "dependencies": {
-    "serverless-optimizer-plugin": "^2.5.1"
+    "3scale": "^0.7.2",
+    "q": "^1.4.1",
+    "redis": "^2.5.3",
+    "redis-commands": "^1.1.0",
+    "request": "^2.71.0",
+    "then-redis": "^2.0.0",
+    "underscore": "*"
   }
 }
